commonize is_az_sql_conn
