+++ tests/conftest.py
@@ -5,12 +5,25 @@ Functions:
 - conn_str: Fixture to get the connection string from environment variables.
 - db_connection: Fixture to create and yield a database connection.
 - cursor: Fixture to create and yield a cursor from the database connection.
+- is_azure_sql_connection: Helper function to detect Azure SQL Database connections.
 """
 
 import pytest
 import os
+import re
 from mssql_python import connect
 import time
+
+
+def is_azure_sql_connection(conn_str):
+    """Helper function to detect if connection string is for Azure SQL Database"""
+    if not conn_str:
+        return False
+    # Check if database.windows.net appears in the Server parameter
+    conn_str_lower = conn_str.lower()
+    # Look for Server= or server= followed by database.windows.net
+    server_match = re.search(r'server\s*=\s*[^;]*database\.windows\.net', conn_str_lower)
+    return server_match is not None
 
 
 def pytest_configure(config):
