+++ tests/test_003_connection.py
@@ -43,18 +43,7 @@ from mssql_python.exceptions import (
 import struct
 from datetime import datetime, timedelta, timezone
 from mssql_python.constants import ConstantsDDBC
-
-
-def is_azure_sql_connection(conn_str):
-    """Helper function to detect if connection string is for Azure SQL Database"""
-    if not conn_str:
-        return False
-    # Check if database.windows.net appears in the Server parameter
-    conn_str_lower = conn_str.lower()
-    # Look for Server= or server= followed by database.windows.net
-    import re
-    server_match = re.search(r'server\s*=\s*[^;]*database\.windows\.net', conn_str_lower)
-    return server_match is not None
+from conftest import is_azure_sql_connection
 
 
 @pytest.fixture(autouse=True)
